hexify nonprintable codes; ignore (for now) elements without plurals
We may have to go back and redo the way we do elements; right now I'm ignoring classes that are not explicitly listed as elements of something, but perhaps a dictionary can be faulty in this regard so I'm never getting a needed plural.

Another possible issue is that I'm not doing anything about bad dictionaries that define the same English term with multiple different codes; I'm just letting appscript mop these up later as it forms its application class.